Module pasteur::middlewares

The middlewares module contents all the public pages

Reexports

use std::prelude::v1::*;

Modules

lang

style

Structs

DesctiptionLang
The DesctiptionLang structure defines an interface to get sentence internationalized.

ShareLang
The ShareLang structure defines and shares the lang by Arc.

ShareStyle
The ShareStyle structure defines and shares the style by Arc.