Module pasteur::middlewares::lang

Reexports

extern crate std;

use std::prelude::v1::*;

use std::io::Read;

use l20n;

use iron;

Structs

DesctiptionLang
The DesctiptionLang structure defines an interface to get sentence internationalized.

Lang

ShareLang
The ShareLang structure defines and shares the lang by Arc.

Type Definitions

Directionary
The Lang structure defines the international table of translation.

Environement